+++ packages/project-input-validator/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/project-input-validator
 
+## 0.3.1
+
+### Patch Changes
+
+-   108336f: Adds prompts for abap on premise systems to odata-service-inquirer
+
 ## 0.3.0
 
 ### Minor Changes